// src/gen/plugin/codegen_pb.ts
export class File extends Message<File> {
	/**
	 * @generated from field: string name = 1;
	 */
	name = "";

	/**
	 * @generated from field: bytes contents = 2;
	 */
	contents = new Uint8Array(0);

	constructor(data?: PartialMessage<File>) {
		super();
		proto3.util.initPartial(data, this);
	}

	static readonly runtime: typeof proto3 = proto3;
	static readonly typeName = "plugin.File";
	static readonly fields: FieldList = proto3.util.newFieldList(() => [
		{ no: 1, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 2, name: "contents", kind: "scalar", T: 12 /* ScalarType.BYTES */ },
	]);

	static fromBinary(
		bytes: Uint8Array,
		options?: Partial<BinaryReadOptions>,
	): File {
		return new File().fromBinary(bytes, options);
	}

	static fromJson(
		jsonValue: JsonValue,
		options?: Partial<JsonReadOptions>,
	): File {
		return new File().fromJson(jsonValue, options);
	}

	static fromJsonString(
		jsonString: string,
		options?: Partial<JsonReadOptions>,
	): File {
		return new File().fromJsonString(jsonString, options);
	}

	static equals(
		a: File | PlainMessage<File> | undefined,
		b: File | PlainMessage<File> | undefined,
	): boolean {
		return proto3.util.equals(File, a, b);
	}
}
// ...
export class Column extends Message<Column> {
	/**
	 * @generated from field: string name = 1;
	 */
	name = "";

	/**
	 * @generated from field: bool not_null = 3;
	 */
	notNull = false;

	/**
	 * @generated from field: bool is_array = 4;
	 */
	isArray = false;

	/**
	 * @generated from field: string comment = 5;
	 */
	comment = "";

	/**
	 * @generated from field: int32 length = 6;
	 */
	length = 0;

	/**
	 * @generated from field: bool is_named_param = 7;
	 */
	isNamedParam = false;

	/**
	 * @generated from field: bool is_func_call = 8;
	 */
	isFuncCall = false;

	/**
	 * XXX: Figure out what PostgreSQL calls `foo.id`
	 *
	 * @generated from field: string scope = 9;
	 */
	scope = "";

	/**
	 * @generated from field: plugin.Identifier table = 10;
	 */
	table?: Identifier;

	/**
	 * @generated from field: string table_alias = 11;
	 */
	tableAlias = "";

	/**
	 * @generated from field: plugin.Identifier type = 12;
	 */
	type?: Identifier;

	/**
	 * @generated from field: bool is_sqlc_slice = 13;
	 */
	isSqlcSlice = false;

	/**
	 * @generated from field: plugin.Identifier embed_table = 14;
	 */
	embedTable?: Identifier;

	/**
	 * @generated from field: string original_name = 15;
	 */
	originalName = "";

	/**
	 * @generated from field: bool unsigned = 16;
	 */
	unsigned = false;

	/**
	 * @generated from field: int32 array_dims = 17;
	 */
	arrayDims = 0;

	constructor(data?: PartialMessage<Column>) {
		super();
		proto3.util.initPartial(data, this);
	}

	static readonly runtime: typeof proto3 = proto3;
	static readonly typeName = "plugin.Column";
	static readonly fields: FieldList = proto3.util.newFieldList(() => [
		{ no: 1, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 3, name: "not_null", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
		{ no: 4, name: "is_array", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
		{ no: 5, name: "comment", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 6, name: "length", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
		{
			no: 7,
			name: "is_named_param",
			kind: "scalar",
			T: 8 /* ScalarType.BOOL */,
		},
		{ no: 8, name: "is_func_call", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
		{ no: 9, name: "scope", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 10, name: "table", kind: "message", T: Identifier },
		{
			no: 11,
			name: "table_alias",
			kind: "scalar",
			T: 9 /* ScalarType.STRING */,
		},
		{ no: 12, name: "type", kind: "message", T: Identifier },
		{
			no: 13,
			name: "is_sqlc_slice",
			kind: "scalar",
			T: 8 /* ScalarType.BOOL */,
		},
		{ no: 14, name: "embed_table", kind: "message", T: Identifier },
		{
			no: 15,
			name: "original_name",
			kind: "scalar",
			T: 9 /* ScalarType.STRING */,
		},
		{ no: 16, name: "unsigned", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
		{ no: 17, name: "array_dims", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
	]);

	static fromBinary(
		bytes: Uint8Array,
		options?: Partial<BinaryReadOptions>,
	): Column {
		return new Column().fromBinary(bytes, options);
	}

	static fromJson(
		jsonValue: JsonValue,
		options?: Partial<JsonReadOptions>,
	): Column {
		return new Column().fromJson(jsonValue, options);
	}

	static fromJsonString(
		jsonString: string,
		options?: Partial<JsonReadOptions>,
	): Column {
		return new Column().fromJsonString(jsonString, options);
	}

	static equals(
		a: Column | PlainMessage<Column> | undefined,
		b: Column | PlainMessage<Column> | undefined,
	): boolean {
		return proto3.util.equals(Column, a, b);
	}
}

/**
 * @generated from message plugin.Query
 */
export class Query extends Message<Query> {
	/**
	 * @generated from field: string text = 1;
	 */
	text = "";

	/**
	 * @generated from field: string name = 2;
	 */
	name = "";

	/**
	 * @generated from field: string cmd = 3;
	 */
	cmd = "";

	/**
	 * @generated from field: repeated plugin.Column columns = 4;
	 */
	columns: Column[] = [];

	/**
	 * @generated from field: repeated plugin.Parameter params = 5 [json_name = "parameters"];
	 */
	params: Parameter[] = [];

	/**
	 * @generated from field: repeated string comments = 6;
	 */
	comments: string[] = [];

	/**
	 * @generated from field: string filename = 7;
	 */
	filename = "";

	/**
	 * @generated from field: plugin.Identifier insert_into_table = 8 [json_name = "insert_into_table"];
	 */
	insertIntoTable?: Identifier;

	constructor(data?: PartialMessage<Query>) {
		super();
		proto3.util.initPartial(data, this);
	}

	static readonly runtime: typeof proto3 = proto3;
	static readonly typeName = "plugin.Query";
	static readonly fields: FieldList = proto3.util.newFieldList(() => [
		{ no: 1, name: "text", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 2, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 3, name: "cmd", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{ no: 4, name: "columns", kind: "message", T: Column, repeated: true },
		{
			no: 5,
			name: "params",
			jsonName: "parameters",
			kind: "message",
			T: Parameter,
			repeated: true,
		},
		{
			no: 6,
			name: "comments",
			kind: "scalar",
			T: 9 /* ScalarType.STRING */,
			repeated: true,
		},
		{ no: 7, name: "filename", kind: "scalar", T: 9 /* ScalarType.STRING */ },
		{
			no: 8,
			name: "insert_into_table",
			jsonName: "insert_into_table",
			kind: "message",
			T: Identifier,
		},
	]);

	static fromBinary(
		bytes: Uint8Array,
		options?: Partial<BinaryReadOptions>,
	): Query {
		return new Query().fromBinary(bytes, options);
	}

	static fromJson(
		jsonValue: JsonValue,
		options?: Partial<JsonReadOptions>,
	): Query {
		return new Query().fromJson(jsonValue, options);
	}

	static fromJsonString(
		jsonString: string,
		options?: Partial<JsonReadOptions>,
	): Query {
		return new Query().fromJsonString(jsonString, options);
	}

	static equals(
		a: Query | PlainMessage<Query> | undefined,
		b: Query | PlainMessage<Query> | undefined,
	): boolean {
		return proto3.util.equals(Query, a, b);
	}
}

/**
 * @generated from message plugin.Parameter
 */
export class Parameter extends Message<Parameter> {
	/**
	 * @generated from field: int32 number = 1;
	 */
	number = 0;

	/**
	 * @generated from field: plugin.Column column = 2;
	 */
	column?: Column;

	constructor(data?: PartialMessage<Parameter>) {
		super();
		proto3.util.initPartial(data, this);
	}

	static readonly runtime: typeof proto3 = proto3;
	static readonly typeName = "plugin.Parameter";
	static readonly fields: FieldList = proto3.util.newFieldList(() => [
		{ no: 1, name: "number", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
		{ no: 2, name: "column", kind: "message", T: Column },
	]);

	static fromBinary(
		bytes: Uint8Array,
		options?: Partial<BinaryReadOptions>,
	): Parameter {
		return new Parameter().fromBinary(bytes, options);
	}

	static fromJson(
		jsonValue: JsonValue,
		options?: Partial<JsonReadOptions>,
	): Parameter {
		return new Parameter().fromJson(jsonValue, options);
	}

	static fromJsonString(
		jsonString: string,
		options?: Partial<JsonReadOptions>,
	): Parameter {
		return new Parameter().fromJsonString(jsonString, options);
	}

	static equals(
		a: Parameter | PlainMessage<Parameter> | undefined,
		b: Parameter | PlainMessage<Parameter> | undefined,
	): boolean {
		return proto3.util.equals(Parameter, a, b);
	}
}

/**
 * @generated from message plugin.GenerateRequest
 */
export class GenerateRequest extends Message<GenerateRequest> {
	/**
	 * @generated from field: plugin.Settings settings = 1;
	 */
	settings?: Settings;

	/**
	 * @generated from field: plugin.Catalog catalog = 2;
	 */
	catalog?: Catalog;

	/**
	 * @generated from field: repeated plugin.Query queries = 3;
	 */
	queries: Query[] = [];

	/**
	 * @generated from field: string sqlc_version = 4 [json_name = "sqlc_version"];
	 */
	sqlcVersion = "";

	/**
	 * @generated from field: bytes plugin_options = 5 [json_name = "plugin_options"];
	 */
	pluginOptions = new Uint8Array(0);

	/**
	 * @generated from field: bytes global_options = 6 [json_name = "global_options"];
	 */
	globalOptions = new Uint8Array(0);

	constructor(data?: PartialMessage<GenerateRequest>) {
		super();
		proto3.util.initPartial(data, this);
	}

	static readonly runtime: typeof proto3 = proto3;
	static readonly typeName = "plugin.GenerateRequest";
	static readonly fields: FieldList = proto3.util.newFieldList(() => [
		{ no: 1, name: "settings", kind: "message", T: Settings },
		{ no: 2, name: "catalog", kind: "message", T: Catalog },
		{ no: 3, name: "queries", kind: "message", T: Query, repeated: true },
		{
			no: 4,
			name: "sqlc_version",
			jsonName: "sqlc_version",
			kind: "scalar",
			T: 9 /* ScalarType.STRING */,
		},
		{
			no: 5,
			name: "plugin_options",
			jsonName: "plugin_options",
			kind: "scalar",
			T: 12 /* ScalarType.BYTES */,
		},
		{
			no: 6,
			name: "global_options",
			jsonName: "global_options",
			kind: "scalar",
			T: 12 /* ScalarType.BYTES */,
		},
	]);

	static fromBinary(
		bytes: Uint8Array,
		options?: Partial<BinaryReadOptions>,
	): GenerateRequest {
		return new GenerateRequest().fromBinary(bytes, options);
	}

	static fromJson(
		jsonValue: JsonValue,
		options?: Partial<JsonReadOptions>,
	): GenerateRequest {
		return new GenerateRequest().fromJson(jsonValue, options);
	}

	static fromJsonString(
		jsonString: string,
		options?: Partial<JsonReadOptions>,
	): GenerateRequest {
		return new GenerateRequest().fromJsonString(jsonString, options);
	}

	static equals(
		a: GenerateRequest | PlainMessage<GenerateRequest> | undefined,
		b: GenerateRequest | PlainMessage<GenerateRequest> | undefined,
	): boolean {
		return proto3.util.equals(GenerateRequest, a, b);
	}
}

/**
 * @generated from message plugin.GenerateResponse
 */
export class GenerateResponse extends Message<GenerateResponse> {
	/**
	 * @generated from field: repeated plugin.File files = 1;
	 */
	files: File[] = [];

	constructor(data?: PartialMessage<GenerateResponse>) {
		super();
		proto3.util.initPartial(data, this);
	}

	static readonly runtime: typeof proto3 = proto3;
	static readonly typeName = "plugin.GenerateResponse";
	static readonly fields: FieldList = proto3.util.newFieldList(() => [
		{ no: 1, name: "files", kind: "message", T: File, repeated: true },
	]);

	static fromBinary(
		bytes: Uint8Array,
		options?: Partial<BinaryReadOptions>,
	): GenerateResponse {
		return new GenerateResponse().fromBinary(bytes, options);
	}

	static fromJson(
		jsonValue: JsonValue,
		options?: Partial<JsonReadOptions>,
	): GenerateResponse {
		return new GenerateResponse().fromJson(jsonValue, options);
	}

	static fromJsonString(
		jsonString: string,
		options?: Partial<JsonReadOptions>,
	): GenerateResponse {
		return new GenerateResponse().fromJsonString(jsonString, options);
	}

	static equals(
		a: GenerateResponse | PlainMessage<GenerateResponse> | undefined,
		b: GenerateResponse | PlainMessage<GenerateResponse> | undefined,
	): boolean {
		return proto3.util.equals(GenerateResponse, a, b);
	}
}